Code for obtaining KPIS for Complete Regions
Based on a folder with csv files of instances of a region being equal/unequal type of evaluation. Unifies and computes average value for region and stores unique instances with its 20 samples calcualtion. Returns also a csv file of all the data of that region_equal/unequal. Visualization graph for evaluation of trends.